+++ project.sh
@@ -16,7 +16,6 @@ $PIP install prefact --upgrade --quiet
 $PIP install vallm --upgrade --quiet
 $PIP install redup --upgrade --quiet
 $PIP install glon --upgrade --quiet
-$PIP install goal --upgrade --quiet
 $PIP install code2logic --upgrade --quiet
 $PIP install code2llm --upgrade --quiet
 #$VENV/bin/code2llm ./ -f toon,evolution,code2logic,project-yaml -o ./project --no-chunk
@@ -33,3 +32,8 @@ $VENV/bin/redup scan . --format toon --output ./project
 #$VENV/bin/vallm batch --parallel .
 $VENV/bin/vallm batch . --recursive --format toon --output ./project
 $VENV/bin/prefact -a -e "examples/**"
+
+
+pip install -U goal
+$PIP install goal --upgrade --quiet
+$VENV/bin/goal -a